extract bench
