+++ blockstore/src/bin/postgres-select1-bencher.rs
@@ -1,10 +1,5 @@
-use std::sync::Arc;
-use std::sync::atomic::AtomicU32;
-use itertools::join;
 use log::info;
-use tokio::join;
-use tokio_postgres::Row;
-use solana_lite_rpc_blockstore::block_stores::postgres::{BlockstorePostgresSessionConfig, PostgresSession};
+use solana_lite_rpc_blockstore::block_stores::postgres::measure_database_roundtrip::measure_select1_roundtrip;
 
 // RUST_LOG=info
 // requires BLOCKSTOREDB_PG_CONFIG
@@ -12,32 +7,9 @@ use solana_lite_rpc_blockstore::block_stores::postgres::{BlockstorePostgresSessi
 async fn main() {
     tracing_subscriber::fmt::init();
 
-    let mut counter = Arc::new(AtomicU32::new(0));
+    let (num_queries, avg_time) = measure_select1_roundtrip().await;
 
-    let mut jh_tasks = vec![];
-    for _session in 0..5 {
-        let counter = counter.clone();
-        let jh = tokio::spawn(async move {
-            let postgres_session = PostgresSession::new_from_env().await.unwrap();
-
-            let started_at = tokio::time::Instant::now();
-            // 100 sequenctial roundtrips
-            for j in 0..1000 {
-                let _result: Row = postgres_session.query_one("SELECT 1", &[]).await.unwrap();
-                counter.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-            }
-
-            (started_at.elapsed())
-
-        });
-        jh_tasks.push(jh);
-    }
-
-    for jh in jh_tasks {
-        let elapsed = jh.await.unwrap();
-        info!("elapsed: {:?}", elapsed);
-    }
-    info!("total counter: {}", counter.load(std::sync::atomic::Ordering::Relaxed));
-
+    info!("total num queris: {}", num_queries);
+    info!("avg roundtrip: {:.2?}", avg_time);
 
 }
