+++ blockstore/src/block_stores/postgres/mod.rs
@@ -1,6 +1,7 @@
 pub mod postgres_block_store_importer;
 pub mod postgres_block_store_query;
 pub mod postgres_block_store_writer;
+pub mod measure_database_roundtrip;
 
 use anyhow::Context;
 pub use postgres_config::BlockstorePostgresSessionConfig;
